@@ -1,49 +1,55 @@
 {
   "name": "visualizer-on-tabs",
-  "version": "0.0.0",
-  "description": "",
+  "version": "1.0.0",
+  "description": "visualizer-on-tabs webpack builder",
+  "type": "module",
+  "exports": {
+    ".": "./main/index.js"
+  },
+  "bin": "bin/build.js",
+  "files": [
+    "main",
+    "src",
+    "bin",
+    "!**/*.test.js"
+  ],
+  "browserslist": [
+    "defaults"
+  ],
   "scripts": {
-    "clean": "shx rm -rf out",
-    "copy-static": "shx mkdir -p out && shx cp -r src/content/* out",
-    "dev": "node bin/build.js --debug=1 --config=./dev.json && npx serve out/",
+    "build": "npx --no-install visualizer-on-tabs --config=./dev.json --outDir=./out",
+    "build:dev": "npx --no-install visualizer-on-tabs --dev=1 --config=./dev.json --outDir=./out",
+    "clean": "rimraf out",
     "eslint": "eslint bin main src",
-    "webpack": "node bin/build.js --debug=1",
-    "watch": "npm run copy-static && npm run webpack -- --watch",
-    "build": "npm run clean && npm run webpack",
-    "test": "jest && npm run eslint"
+    "prettier": "prettier --check ./",
+    "prettier-write": "prettier --write ./",
+    "serve": "npx serve out/",
+    "test": "npm run test-only && npm run eslint && npm run prettier",
+    "test-only": "node --test"
   },
   "dependencies": {
-    "@babel/core": "^7.24.4",
-    "@babel/preset-env": "^7.24.4",
-    "@babel/preset-react": "^7.24.1",
-    "babel-loader": "^9.1.3",
+    "@babel/core": "^7.28.3",
+    "@babel/preset-env": "^7.28.3",
+    "@babel/preset-react": "^7.27.1",
+    "babel-loader": "^10.0.0",
     "bootstrap": "^5.3.3",
-    "fs-extra": "^7.0.1",
-    "iframe-bridge": "^1.1.0",
-    "js-yaml": "^3.12.1",
+    "iframe-bridge": "^3.0.2",
     "lockr": "^0.8.5",
     "lodash": "^4.17.21",
     "minimist": "^1.2.6",
     "react": "^18.3.1",
-    "react-bootstrap": "^2.10.2",
+    "react-bootstrap": "^3.0.0-beta.3",
     "react-dom": "^18.3.1",
     "react-visualizer": "^3.0.1",
-    "superagent": "^4.1.0",
     "webpack": "^5.91.0"
   },
-  "main": "main/index.js",
   "devDependencies": {
-    "babel-eslint": "^10.0.1",
-    "eslint": "^5.12.0",
-    "eslint-config-cheminfo": "^1.20.0",
-    "eslint-config-neptune-react": "^2.6.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.1.3",
-    "eslint-plugin-react": "^7.12.4",
-    "jest": "^29.7.0",
-    "shx": "^0.3.2"
+    "eslint": "^9.34.0",
+    "eslint-config-zakodium": "^16.0.0",
+    "prettier": "^3.6.2",
+    "rimraf": "^6.0.1"
   },
   "volta": {
-    "node": "20.12.2"
+    "node": "24.7.0"
   }
 }
