@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import App from './components/App';
+import App from './components/App.js';
 
 const loc = window.location;
 let hash = loc.hash.slice(1);
@@ -11,8 +11,10 @@ if (!hash) {
 
 const props = {
   origin: loc.origin,
-  path: hash
+  path: hash,
 };
 
 const element = React.createElement(App, props);
-ReactDOM.createRoot(document.getElementById('visualizer-on-tabs')).render(element);
+ReactDOM.createRoot(document.getElementById('visualizer-on-tabs')).render(
+  element,
+);
