@@ -1,6 +1,6 @@
 import lockr from 'lockr';
 
-import { version } from './constants';
+import { version } from './constants.js';
 
 const LOCAL_STORAGE_TAB_DATA = 'vweb-';
 const LOCAL_STORAGE_TAB_IDS = 'vweb1-tab-ids';
@@ -12,28 +12,28 @@ function isVersionOK(v) {
   return v === undefined || v === version;
 }
 
-storage.save = function (tabId, data) {
+export function save(tabId, data) {
   if (!tabId) return;
   lockr.sadd(LOCAL_STORAGE_TAB_IDS, tabId);
-  var key = LOCAL_STORAGE_TAB_DATA + tabId;
+  let key = LOCAL_STORAGE_TAB_DATA + tabId;
   data.version = version;
   lockr.set(key, data);
-};
+}
 
-storage.saveSelected = function (tabId) {
+export function saveSelected(tabId) {
   if (!tabId) return;
   lockr.set(LOCAL_STORAGE_LAST_TAB, tabId);
-};
+}
 
-storage.getSelected = function () {
+export function getSelected() {
   return lockr.get(LOCAL_STORAGE_LAST_TAB);
-};
+}
 
-storage.load = function () {
-  var ids = lockr.smembers(LOCAL_STORAGE_TAB_IDS);
+export function load() {
+  let ids = lockr.smembers(LOCAL_STORAGE_TAB_IDS);
   if (!ids) return [];
 
-  var data = ids.map((id) => {
+  let data = ids.map((id) => {
     return lockr.get(LOCAL_STORAGE_TAB_DATA + id);
   });
 
@@ -45,19 +45,19 @@ storage.load = function () {
 
   data = data.filter((entry) => isVersionOK(entry.version));
 
-  data.sort(function (a, b) {
-    var idxA = ids.indexOf(a.id);
-    var idxB = ids.indexOf(b.id);
+  data.sort(function sortData(a, b) {
+    let idxA = ids.indexOf(a.id);
+    let idxB = ids.indexOf(b.id);
     if (idxA < idxB) return -1;
     else if (idxB < idxA) return 1;
     else return 0;
   });
   return data;
-};
+}
 
-storage.remove = function (id) {
+export function remove(id) {
   lockr.srem(LOCAL_STORAGE_TAB_IDS, id);
   lockr.rm(LOCAL_STORAGE_TAB_DATA + id);
-};
+}
 
 export default storage;
